# 🚀 Add Punch Hole Caching System for Enhanced Template Performance