--- a/ast/hole_statment.go
+++ b/ast/hole_statment.go
@@ -1,0 +1,13 @@
+package ast
+
+import "strings"
+
+type HoleStatement struct {
+	TokenAble
+	Statements string
+}
+
+func (h *HoleStatement) statementNode() {}
+func (h *HoleStatement) String() string {
+	return strings.TrimSpace(h.Statements)
+}