--- a/partial_helper.go
+++ b/partial_helper.go
@@ -14,7 +14,6 @@ func PartialHelper(name string, data map[string]interface{}, help HelperContext)
 	if help.Context == nil {
 		return "", fmt.Errorf("invalid context. abort")
 	}
-
 	help.Context = help.New()
 	for k, v := range data {
 		help.Set(k, v)
@@ -30,11 +29,10 @@ func PartialHelper(name string, data map[string]interface{}, help HelperContext)
 	if part, err = pf(name); err != nil {
 		return "", err
 	}
-
+	help.Context.Set(TemplateFileKey, name)
 	if part, err = Render(part, help.Context); err != nil {
 		return "", err
 	}
-
 	if ct, ok := help.Value("contentType").(string); ok {
 		ext := filepath.Ext(name)
 		if strings.Contains(ct, "javascript") && ext != ".js" && ext != "" {