--- a/plush.go
+++ b/plush.go
@@ -1,11 +1,14 @@
 package plush
 
 import (
+	"errors"
 	"fmt"
 	"html/template"
 	"io"
 	"io/ioutil"
+	"strings"
 	"sync"
+	"time"
 
 	"github.com/gobuffalo/plush/v5/helpers/hctx"
 )
@@ -22,67 +25,151 @@ import (
 */
 var DefaultTimeFormat = "January 02, 2006 15:04:05 -0700"
 
-var CacheEnabled bool
-var cache = map[string]*Template{}
-var moot = &sync.Mutex{}
+var holeTemplateFileKey = "hole_render_key_plush_" + fmt.Sprintf("%d", time.Now().UnixNano())
+var TemplateFileKey = "renderd_file_name_key_plush_" + fmt.Sprintf("%d", time.Now().UnixNano())
+var cacheEnabled bool
 
-func CacheSet(key string, t *Template) {
-	moot.Lock()
-	defer moot.Unlock()
+var templateCacheBackend TemplateCache
 
-	cache[key] = t
+func PlushCacheSetup(ts TemplateCache) {
+	cacheEnabled = true
+	templateCacheBackend = ts
 }
 
 // BuffaloRenderer implements the render.TemplateEngine interface allowing velvet to be used as a template engine
 // for Buffalo
 func BuffaloRenderer(input string, data map[string]interface{}, helpers map[string]interface{}) (string, error) {
-	t, err := Parse(input)
-	if err != nil {
-		return "", err
+	if data == nil {
+		data = make(map[string]interface{})
 	}
-
 	for k, v := range helpers {
 		data[k] = v
 	}
-	gs := NewContextWith(data)
+	ctx := NewContextWith(data)
 	defer func() {
-		for k := range gs.data.localInterner.stringToID {
-			data[k] = gs.Value(k)
+		if data != nil {
+			for k := range ctx.data.localInterner.stringToID {
+				data[k] = ctx.Value(k)
+			}
 		}
 	}()
-	return t.Exec(gs)
+	return Render(input, ctx)
 }
 
 // Parse an input string and return a Template, and caches the parsed template.
-func Parse(input string) (*Template, error) {
-	if !CacheEnabled {
-		return NewTemplate(input)
+func Parse(input ...string) (*Template, error) {
+	if templateCacheBackend == nil || !cacheEnabled || len(input) == 1 || len(input) > 2 {
+		return NewTemplate(input[0])
 	}
 
-	moot.Lock()
-	defer moot.Unlock()
-
-	t, ok := cache[input]
-	if ok {
-		return t, nil
+	filename := input[1]
+	if filename != "" && templateCacheBackend != nil {
+		t, ok := templateCacheBackend.Get(filename)
+		if ok {
+			return t, nil
+		}
 	}
 
-	t, err := NewTemplate(input)
+	t, err := NewTemplate(input[0])
 	if err != nil {
 		return t, err
 	}
-
-	cache[input] = t
 	return t, nil
 }
 
-// Render a string using the given the context.
+// Render a string using the given context.
 func Render(input string, ctx hctx.Context) (string, error) {
-	t, err := Parse(input)
+	var filename string
+
+	// Extract filename from context if we're not in a hole rendering pass.
+	// The filename is used for template caching - only main templates (not holes) should use cache.
+	if ctx.Value(holeTemplateFileKey) == nil && ctx.Value(TemplateFileKey) != nil {
+		filename, _ = ctx.Value(TemplateFileKey).(string)
+	}
+
+	// Try to render from cache if conditions are met:
+	// - Not in hole rendering pass (prevents infinite recursion)
+	// - Cache is enabled and backend is available
+	// - Template has a filename for cache key
+	if ctx.Value(holeTemplateFileKey) == nil {
+		cacheT, cacheErr := renderFromCache(filename, ctx)
+		if cacheErr == nil {
+			return cacheT, nil
+		}
+	}
+
+	// Parse template (with caching if enabled)
+	t, err := Parse(input, filename)
 	if err != nil {
 		return "", err
 	}
-	return t.Exec(ctx)
+
+	// Execute template to get skeleton with hole markers
+	s, err := t.Exec(ctx)
+	if err != nil {
+		return "", err
+	}
+	t.skeleton = s
+
+	// Cache the template after successful execution (deferred to ensure caching even if hole rendering fails)
+	defer func() {
+		if cacheEnabled && templateCacheBackend != nil && filename != "" {
+			t.punchHole = holesCopy(t.punchHole)
+			templateCacheBackend.Set(filename, t)
+		}
+	}()
+
+	// If we have holes and this is the main render pass (not hole rendering),
+	// render holes concurrently and fill them into the skeleton
+	if ctx.Value(holeTemplateFileKey) == nil && len(t.punchHole) > 0 {
+		hc := renderHolesConcurrently(t.punchHole, ctx)
+		return fillHoles(s, hc)
+	}
+
+	// Return skeleton as-is (either no holes or we're in hole rendering pass)
+	return s, nil
+}
+
+// fillHoles replaces all markers in the rendered string with their rendered content using stored positions.
+func fillHoles(rendered string, holes []HoleMarker) (string, error) {
+
+	var sb strings.Builder
+	last := 0
+	for _, pos := range holes {
+		if pos.err != nil {
+			return "", pos.err
+		}
+		sb.WriteString(rendered[last:pos.start])
+		sb.WriteString(pos.content)
+		last = pos.end
+	}
+	sb.WriteString(rendered[last:])
+	return sb.String(), nil
+}
+
+func renderHolesConcurrently(holes []HoleMarker, ctx hctx.Context) []HoleMarker {
+	var wg sync.WaitGroup
+	wg.Add(len(holes))
+	// Set a special key in the context to indicate we are rendering holes
+	// This prevents infinite recursion if a hole template itself has holes
+	// or if the hole template is cached
+	defer ctx.Update(holeTemplateFileKey, nil)
+	ctx.Set(holeTemplateFileKey, true)
+	for k, hole := range holes {
+
+		go func(k int, h HoleMarker) {
+			defer wg.Done()
+			// Render the hole's content (this could be a Plush template or any logic)
+			content, err := Render(h.input, ctx)
+			if err != nil {
+				holes[k].err = err
+				return
+			}
+			holes[k].content = content
+		}(k, hole)
+	}
+	wg.Wait()
+	return holes
 }
 
 func RenderR(input io.Reader, ctx hctx.Context) (string, error) {
@@ -116,4 +203,40 @@ type interfaceable interface {
 // HTMLer generates HTML source
 type HTMLer interface {
 	HTML() template.HTML
+}
+
+func holesCopy(holes []HoleMarker) []HoleMarker {
+	holesCopy := make([]HoleMarker, len(holes))
+	for i := range holes {
+		holesCopy[i] = holes[i]
+		holesCopy[i].content = ""
+		holesCopy[i].err = nil
+	}
+	return holesCopy
+}
+
+// This function tries to get the template from the cache if it exists
+// It only works if we are not in a hole rendering pass, and if we have a filename
+// and if cache is enabled and if we have a cache backend
+// If we are in a hole rendering pass, we should not use the cache.
+// If we are in the first pass, and there is a filename, we should use the cache.
+// If there is no filename, we should not use the cache.
+// If cache is disabled, we should not use the cache.
+// If there is no templateCacheBackend, we should not use the cache.
+func renderFromCache(filename string, ctx hctx.Context) (string, error) {
+
+	if filename != "" && cacheEnabled && templateCacheBackend != nil && ctx.Value(holeTemplateFileKey) == nil {
+		if filename != "" {
+			inCacheTemplate, inCache := templateCacheBackend.Get(filename)
+			if inCache &&
+				inCacheTemplate != nil &&
+				inCacheTemplate.skeleton != "" &&
+				len(inCacheTemplate.punchHole) > 0 {
+				hc := holesCopy(inCacheTemplate.punchHole)
+				hc = renderHolesConcurrently(hc, ctx)
+				return fillHoles(inCacheTemplate.skeleton, hc)
+			}
+		}
+	}
+	return "", errors.New("no cached template found")
 }