--- a/plush.go
+++ b/plush.go
@@ -150,17 +150,26 @@ func fillHoles(rendered string, holes []HoleMarker) (string, error) {
 func renderHolesConcurrently(holes []HoleMarker, ctx hctx.Context) []HoleMarker {
 	var wg sync.WaitGroup
 	wg.Add(len(holes))
-	// Set a special key in the context to indicate we are rendering holes
-	// This prevents infinite recursion if a hole template itself has holes
-	// or if the hole template is cached
-	defer ctx.Update(holeTemplateFileKey, nil)
-	ctx.Set(holeTemplateFileKey, true)
+
+	octx := ctx
+	defer func() {
+		ctx = octx
+	}()
+
 	for k, hole := range holes {
 
 		go func(k int, h HoleMarker) {
 			defer wg.Done()
+			// Create a new isolated context for each hole to prevent race conditions
+			// Each hole gets its own copy of the context, preventing concurrent modifications
+			holeCtx := ctx.New()
+			// Set a special key in the context to indicate we are rendering holes
+			// This prevents infinite recursion if a hole template itself has holes
+			// or if the hole template is cached
+			holeCtx.Set(holeTemplateFileKey, true)
+
 			// Render the hole's content (this could be a Plush template or any logic)
-			content, err := Render(h.input, ctx)
+			content, err := Render(h.input, holeCtx)
 			if err != nil {
 				holes[k].err = err
 				return