--- a/template.go
+++ b/template.go
@@ -11,9 +11,10 @@ import (
 // to evaluate and render the input.
 type Template struct {
 	Input     string
-	program   *ast.Program
-	punchHole []HoleMarker
-	skeleton  string
+	Program   *ast.Program
+	PunchHole []HoleMarker
+	Skeleton  string
+	IsCache   bool
 }
 
 // NewTemplate from the input string. Adds all of the
@@ -36,7 +37,7 @@ func NewTemplate(input string) (*Template, error) {
 // as a successful result is cached and is used on subsequent
 // uses.
 func (t *Template) Parse() error {
-	if t.program != nil {
+	if t.Program != nil {
 		return nil
 	}
 
@@ -45,7 +46,7 @@ func (t *Template) Parse() error {
 		return err
 	}
 
-	t.program = program
+	t.Program = program
 	return nil
 }
 
@@ -58,7 +59,7 @@ func (t *Template) Exec(ctx hctx.Context) (string, []HoleMarker, error) {
 
 	ev := compiler{
 		ctx:     ctx,
-		program: t.program,
+		program: t.Program,
 	}
 
 	s, err := ev.compile()
@@ -69,7 +70,7 @@ func (t *Template) Exec(ctx hctx.Context) (string, []HoleMarker, error) {
 func (t *Template) Clone() *Template {
 	t2 := &Template{
 		Input:   t.Input,
-		program: t.program,
+		Program: t.Program,
 	}
 	return t2
 }